refactor(core): extract standalone modules from database/core.rs
Splits four self-contained pieces out of the 5,338-line core.rs into
sibling modules so that working on each one no longer requires loading
the whole file:

- conversions.rs (353 lines): timestamp/date/tag-blob helpers and the
  TaskStatus/TaskType from_i32 mappers
- pool.rs (339 lines): DatabasePoolConfig, SqliteOptimizations,
  PoolHealthStatus, PoolMetrics, ComprehensiveHealthStatus
- stats.rs (58 lines): DatabaseStats
- path_discovery.rs (152 lines): get_default_database_path and the
  ThingsData-* directory scan

Tests for each piece move with the code into a #[cfg(test)] mod tests
inside each new file. core.rs drops from 5,338 to 4,479 lines.

This is a mechanical move with zero behavior change. Public API
re-exports from `database::*` are preserved via mod.rs, so neither
lib.rs nor external test consumers (write_operations_tests.rs,
date_handling_tests.rs) need to change.

The giant `impl ThingsDatabase` block remains in core.rs for now;
splitting it by domain (queries/, mutations/) is the next stage of the
broader plan in ~/.claude/plans/lately-it-seems-like-lexical-ullman.md.

Verification: cargo test --workspace --all-features yields 1864 passed,
0 failed; cargo clippy -p things3-core --features advanced-queries
-- -D warnings clean.
