+++ libs/things3-core/src/database/core.rs
@@ -2,8 +2,14 @@
 use crate::models::TaskFilters;
 use crate::{
     database::{
+        conversions::{naive_date_to_things_timestamp, safe_timestamp_convert},
         mappers::{map_project_row, map_task_row},
+        pool::{
+            ComprehensiveHealthStatus, DatabasePoolConfig, PoolHealthStatus, PoolMetrics,
+            SqliteOptimizations,
+        },
         query_builders::TaskUpdateBuilder,
+        stats::DatabaseStats,
         validators,
     },
     error::{Result as ThingsResult, ThingsError},
@@ -13,215 +19,11 @@ use crate::{
     },
 };
 use chrono::{DateTime, NaiveDate, Utc};
-use serde::{Deserialize, Serialize};
 use sqlx::{pool::PoolOptions, Row, SqlitePool};
-use std::path::{Path, PathBuf};
-use std::time::Duration;
+use std::path::Path;
 use tracing::{debug, error, info, instrument};
 #[cfg(any(feature = "advanced-queries", feature = "batch-operations"))]
 use uuid::Uuid;
-
-/// Convert f64 timestamp to i64 safely
-pub(crate) fn safe_timestamp_convert(ts_f64: f64) -> i64 {
-    // Use try_from to avoid clippy warnings about casting
-    if ts_f64.is_finite() && ts_f64 >= 0.0 {
-        // Use a reasonable upper bound for timestamps (year 2100)
-        let max_timestamp = 4_102_444_800_f64; // 2100-01-01 00:00:00 UTC
-        if ts_f64 <= max_timestamp {
-            // Convert via string to avoid precision loss warnings
-            let ts_str = format!("{:.0}", ts_f64.trunc());
-            ts_str.parse::<i64>().unwrap_or(0)
-        } else {
-            0 // Use epoch if too large
-        }
-    } else {
-        0 // Use epoch if invalid
-    }
-}
-
-/// Convert Things 3 date value (seconds since 2001-01-01) to NaiveDate
-pub(crate) fn things_date_to_naive_date(seconds_since_2001: i64) -> Option<chrono::NaiveDate> {
-    use chrono::{TimeZone, Utc};
-
-    if seconds_since_2001 <= 0 {
-        return None;
-    }
-
-    // Base date: 2001-01-01 00:00:00 UTC
-    let base_date = Utc.with_ymd_and_hms(2001, 1, 1, 0, 0, 0).single().unwrap();
-
-    // Add seconds to get the actual date
-    let date_time = base_date + chrono::Duration::seconds(seconds_since_2001);
-
-    Some(date_time.date_naive())
-}
-
-/// Convert NaiveDate to Things 3 timestamp (seconds since 2001-01-01)
-pub fn naive_date_to_things_timestamp(date: NaiveDate) -> i64 {
-    use chrono::{NaiveTime, TimeZone, Utc};
-
-    // Base date: 2001-01-01 00:00:00 UTC
-    let base_date = Utc.with_ymd_and_hms(2001, 1, 1, 0, 0, 0).single().unwrap();
-
-    // Convert NaiveDate to DateTime at midnight UTC
-    let date_time = date
-        .and_time(NaiveTime::from_hms_opt(0, 0, 0).unwrap())
-        .and_local_timezone(Utc)
-        .single()
-        .unwrap();
-
-    // Calculate seconds difference
-    date_time.timestamp() - base_date.timestamp()
-}
-
-/// Serialize tags to Things 3 binary format
-/// Note: This is a simplified implementation using JSON
-/// The actual Things 3 binary format is proprietary
-pub fn serialize_tags_to_blob(tags: &[String]) -> ThingsResult<Vec<u8>> {
-    serde_json::to_vec(tags)
-        .map_err(|e| ThingsError::unknown(format!("Failed to serialize tags: {e}")))
-}
-
-/// Deserialize tags from Things 3 binary format
-pub fn deserialize_tags_from_blob(blob: &[u8]) -> ThingsResult<Vec<String>> {
-    if blob.is_empty() {
-        return Ok(Vec::new());
-    }
-    serde_json::from_slice(blob)
-        .map_err(|e| ThingsError::unknown(format!("Failed to deserialize tags: {e}")))
-}
-
-impl TaskStatus {
-    fn from_i32(value: i32) -> Option<Self> {
-        match value {
-            0 => Some(TaskStatus::Incomplete),
-            2 => Some(TaskStatus::Canceled),
-            3 => Some(TaskStatus::Completed),
-            _ => None,
-        }
-    }
-}
-
-impl TaskType {
-    fn from_i32(value: i32) -> Option<Self> {
-        match value {
-            0 => Some(TaskType::Todo),
-            1 => Some(TaskType::Project),
-            2 => Some(TaskType::Heading),
-            3 => Some(TaskType::Area),
-            _ => None,
-        }
-    }
-}
-
-/// Database connection pool configuration for optimal performance
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DatabasePoolConfig {
-    /// Maximum number of connections in the pool
-    pub max_connections: u32,
-    /// Minimum number of connections in the pool
-    pub min_connections: u32,
-    /// Connection timeout
-    pub connect_timeout: Duration,
-    /// Idle timeout for connections
-    pub idle_timeout: Duration,
-    /// Maximum lifetime of a connection
-    pub max_lifetime: Duration,
-    /// Test connections before use
-    pub test_before_acquire: bool,
-    /// SQLite-specific optimizations
-    pub sqlite_optimizations: SqliteOptimizations,
-}
-
-/// SQLite-specific optimization settings
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SqliteOptimizations {
-    /// Enable WAL mode for better concurrency
-    pub enable_wal_mode: bool,
-    /// Set synchronous mode (NORMAL, FULL, OFF)
-    pub synchronous_mode: String,
-    /// Cache size in pages (negative = KB)
-    pub cache_size: i32,
-    /// Enable foreign key constraints
-    pub enable_foreign_keys: bool,
-    /// Set journal mode
-    pub journal_mode: String,
-    /// Set temp store (MEMORY, FILE, DEFAULT)
-    pub temp_store: String,
-    /// Set mmap size for better performance
-    pub mmap_size: i64,
-    /// Enable query planner optimizations
-    pub enable_query_planner: bool,
-}
-
-impl Default for DatabasePoolConfig {
-    fn default() -> Self {
-        Self {
-            max_connections: 10,
-            min_connections: 1,
-            connect_timeout: Duration::from_secs(30),
-            idle_timeout: Duration::from_secs(600), // 10 minutes
-            max_lifetime: Duration::from_secs(1800), // 30 minutes
-            test_before_acquire: true,
-            sqlite_optimizations: SqliteOptimizations::default(),
-        }
-    }
-}
-
-impl Default for SqliteOptimizations {
-    fn default() -> Self {
-        Self {
-            enable_wal_mode: true,
-            synchronous_mode: "NORMAL".to_string(),
-            cache_size: -20000, // 20MB cache
-            enable_foreign_keys: true,
-            journal_mode: "WAL".to_string(),
-            temp_store: "MEMORY".to_string(),
-            mmap_size: 268_435_456, // 256MB
-            enable_query_planner: true,
-        }
-    }
-}
-
-/// Connection pool health status
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PoolHealthStatus {
-    pub is_healthy: bool,
-    pub pool_size: u32,
-    pub active_connections: u32,
-    pub idle_connections: u32,
-    pub max_connections: u32,
-    pub min_connections: u32,
-    pub connection_timeout: Duration,
-    pub idle_timeout: Option<Duration>,
-    pub max_lifetime: Option<Duration>,
-}
-
-/// Detailed connection pool metrics
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PoolMetrics {
-    pub pool_size: u32,
-    pub active_connections: u32,
-    pub idle_connections: u32,
-    pub max_connections: u32,
-    pub min_connections: u32,
-    pub utilization_percentage: f64,
-    pub is_healthy: bool,
-    pub response_time_ms: u64,
-    pub connection_timeout: Duration,
-    pub idle_timeout: Option<Duration>,
-    pub max_lifetime: Option<Duration>,
-}
-
-/// Comprehensive health status including pool and database
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ComprehensiveHealthStatus {
-    pub overall_healthy: bool,
-    pub pool_health: PoolHealthStatus,
-    pub pool_metrics: PoolMetrics,
-    pub database_stats: DatabaseStats,
-    pub timestamp: DateTime<Utc>,
-}
 
 /// SQLx-based database implementation for Things 3 data
 /// This provides async, Send + Sync compatible database access
@@ -3650,97 +3452,10 @@ impl ThingsDatabase {
     }
 }
 
-/// Database statistics
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DatabaseStats {
-    pub task_count: u64,
-    pub project_count: u64,
-    pub area_count: u64,
-}
-
-impl DatabaseStats {
-    #[must_use]
-    pub fn total_items(&self) -> u64 {
-        self.task_count + self.project_count + self.area_count
-    }
-}
-
-/// Things 3 group container directory under the user's `Library`.
-const THINGS_GROUP_CONTAINER: &str =
-    "Library/Group Containers/JLMPQHK86H.com.culturedcode.ThingsMac";
-
-/// Path inside a `ThingsData-XXXXX` directory that points at the SQLite file.
-const THINGS_DB_RELATIVE: &str = "Things Database.thingsdatabase/main.sqlite";
-
-/// Get the default Things 3 database path.
-///
-/// The 4-character suffix on `ThingsData-XXXXX` varies per install (App Store
-/// vs. direct purchase, possibly tied to iCloud account), so this function
-/// scans the group container for any `ThingsData-*` directory containing a
-/// real database file. If multiple candidates exist, the one whose
-/// `main.sqlite` was modified most recently wins. When no candidate is found
-/// the function falls back to the historical literal `ThingsData-0Z0Z2` path
-/// — callers downstream surface a clean "file not found" error in that case.
-///
-/// # Examples
-///
-/// ```
-/// use things3_core::get_default_database_path;
-///
-/// let path = get_default_database_path();
-/// assert!(!path.to_string_lossy().is_empty());
-/// assert!(path.to_string_lossy().contains("Library"));
-/// ```
-#[must_use]
-pub fn get_default_database_path() -> PathBuf {
-    let home = std::env::var("HOME").unwrap_or_else(|_| "~".to_string());
-    let group_container = PathBuf::from(&home).join(THINGS_GROUP_CONTAINER);
-
-    if let Some(found) = discover_things_database(&group_container) {
-        return found;
-    }
-
-    group_container
-        .join("ThingsData-0Z0Z2")
-        .join(THINGS_DB_RELATIVE)
-}
-
-/// Scan `group_container` for `ThingsData-*/Things Database.thingsdatabase/main.sqlite`
-/// and return the most-recently-modified candidate, if any.
-fn discover_things_database(group_container: &std::path::Path) -> Option<PathBuf> {
-    let entries = std::fs::read_dir(group_container).ok()?;
-
-    let mut best: Option<(PathBuf, std::time::SystemTime)> = None;
-    for entry in entries.flatten() {
-        let name = entry.file_name();
-        let Some(name_str) = name.to_str() else {
-            continue;
-        };
-        if !name_str.starts_with("ThingsData-") {
-            continue;
-        }
-
-        let candidate = entry.path().join(THINGS_DB_RELATIVE);
-        let Ok(meta) = std::fs::metadata(&candidate) else {
-            continue;
-        };
-        if !meta.is_file() {
-            continue;
-        }
-        let mtime = meta.modified().unwrap_or(std::time::UNIX_EPOCH);
-
-        match &best {
-            Some((_, best_mtime)) if mtime <= *best_mtime => {}
-            _ => best = Some((candidate, mtime)),
-        }
-    }
-
-    best.map(|(path, _)| path)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::time::Duration;
     use tempfile::{NamedTempFile, TempDir};
 
     #[tokio::test]
@@ -3758,389 +3473,6 @@ mod tests {
     async fn test_connection_string() {
         let result = super::ThingsDatabase::from_connection_string("sqlite::memory:").await;
         assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_task_status_from_i32() {
-        assert_eq!(TaskStatus::from_i32(0), Some(TaskStatus::Incomplete));
-        assert_eq!(TaskStatus::from_i32(1), None); // unused in real Things 3
-        assert_eq!(TaskStatus::from_i32(2), Some(TaskStatus::Canceled));
-        assert_eq!(TaskStatus::from_i32(3), Some(TaskStatus::Completed));
-        assert_eq!(TaskStatus::from_i32(4), None);
-        assert_eq!(TaskStatus::from_i32(-1), None);
-    }
-
-    #[test]
-    fn test_task_type_from_i32() {
-        assert_eq!(TaskType::from_i32(0), Some(TaskType::Todo));
-        assert_eq!(TaskType::from_i32(1), Some(TaskType::Project));
-        assert_eq!(TaskType::from_i32(2), Some(TaskType::Heading));
-        assert_eq!(TaskType::from_i32(3), Some(TaskType::Area));
-        assert_eq!(TaskType::from_i32(4), None);
-        assert_eq!(TaskType::from_i32(-1), None);
-    }
-
-    #[test]
-    fn test_database_stats_total_items() {
-        let stats = DatabaseStats {
-            task_count: 10,
-            project_count: 5,
-            area_count: 3,
-        };
-        assert_eq!(stats.total_items(), 18);
-
-        let empty_stats = DatabaseStats {
-            task_count: 0,
-            project_count: 0,
-            area_count: 0,
-        };
-        assert_eq!(empty_stats.total_items(), 0);
-    }
-
-    #[test]
-    fn test_database_pool_config_default() {
-        let config = DatabasePoolConfig::default();
-        assert_eq!(config.max_connections, 10);
-        assert_eq!(config.min_connections, 1);
-        assert_eq!(config.connect_timeout, Duration::from_secs(30));
-        assert_eq!(config.idle_timeout, Duration::from_secs(600));
-        assert_eq!(config.max_lifetime, Duration::from_secs(1800));
-        assert!(config.test_before_acquire);
-    }
-
-    #[test]
-    fn test_sqlite_optimizations_default() {
-        let opts = SqliteOptimizations::default();
-        assert!(opts.enable_wal_mode);
-        assert_eq!(opts.cache_size, -20000);
-        assert_eq!(opts.synchronous_mode, "NORMAL".to_string());
-        assert_eq!(opts.temp_store, "MEMORY".to_string());
-        assert_eq!(opts.journal_mode, "WAL".to_string());
-        assert_eq!(opts.mmap_size, 268_435_456);
-        assert!(opts.enable_foreign_keys);
-        assert!(opts.enable_query_planner);
-    }
-
-    #[test]
-    fn test_pool_health_status_creation() {
-        let status = PoolHealthStatus {
-            is_healthy: true,
-            pool_size: 8,
-            active_connections: 5,
-            idle_connections: 3,
-            max_connections: 10,
-            min_connections: 1,
-            connection_timeout: Duration::from_secs(30),
-            idle_timeout: Some(Duration::from_secs(600)),
-            max_lifetime: Some(Duration::from_secs(1800)),
-        };
-        assert!(status.is_healthy);
-        assert_eq!(status.active_connections, 5);
-        assert_eq!(status.idle_connections, 3);
-        assert_eq!(status.pool_size, 8);
-    }
-
-    #[test]
-    fn test_pool_metrics_creation() {
-        let metrics = PoolMetrics {
-            pool_size: 8,
-            active_connections: 5,
-            idle_connections: 3,
-            max_connections: 10,
-            min_connections: 1,
-            utilization_percentage: 80.0,
-            is_healthy: true,
-            response_time_ms: 50,
-            connection_timeout: Duration::from_secs(30),
-            idle_timeout: Some(Duration::from_secs(600)),
-            max_lifetime: Some(Duration::from_secs(1800)),
-        };
-        assert!(metrics.is_healthy);
-        assert_eq!(metrics.pool_size, 8);
-        assert_eq!(metrics.active_connections, 5);
-        assert_eq!(metrics.idle_connections, 3);
-        assert!((metrics.utilization_percentage - 80.0).abs() < f64::EPSILON);
-        assert_eq!(metrics.response_time_ms, 50);
-    }
-
-    #[test]
-    fn test_comprehensive_health_status_creation() {
-        let pool_health = PoolHealthStatus {
-            is_healthy: true,
-            pool_size: 8,
-            active_connections: 5,
-            idle_connections: 3,
-            max_connections: 10,
-            min_connections: 1,
-            connection_timeout: Duration::from_secs(30),
-            idle_timeout: Some(Duration::from_secs(600)),
-            max_lifetime: Some(Duration::from_secs(1800)),
-        };
-
-        let pool_metrics = PoolMetrics {
-            pool_size: 8,
-            active_connections: 5,
-            idle_connections: 3,
-            max_connections: 10,
-            min_connections: 1,
-            utilization_percentage: 80.0,
-            is_healthy: true,
-            response_time_ms: 50,
-            connection_timeout: Duration::from_secs(30),
-            idle_timeout: Some(Duration::from_secs(600)),
-            max_lifetime: Some(Duration::from_secs(1800)),
-        };
-
-        let db_stats = DatabaseStats {
-            task_count: 50,
-            project_count: 10,
-            area_count: 5,
-        };
-
-        let health_status = ComprehensiveHealthStatus {
-            overall_healthy: true,
-            pool_health,
-            pool_metrics,
-            database_stats: db_stats,
-            timestamp: Utc::now(),
-        };
-
-        assert!(health_status.overall_healthy);
-        assert_eq!(health_status.database_stats.total_items(), 65);
-    }
-
-    #[test]
-    fn test_safe_timestamp_convert_edge_cases() {
-        // Test normal timestamp
-        assert_eq!(safe_timestamp_convert(1_609_459_200.0), 1_609_459_200); // 2021-01-01
-
-        // Test zero
-        assert_eq!(safe_timestamp_convert(0.0), 0);
-
-        // Test negative (should return 0)
-        assert_eq!(safe_timestamp_convert(-1.0), 0);
-
-        // Test infinity (should return 0)
-        assert_eq!(safe_timestamp_convert(f64::INFINITY), 0);
-
-        // Test NaN (should return 0)
-        assert_eq!(safe_timestamp_convert(f64::NAN), 0);
-
-        // Test very large timestamp (should return 0)
-        assert_eq!(safe_timestamp_convert(5_000_000_000.0), 0);
-
-        // Test max valid timestamp
-        let max_timestamp = 4_102_444_800_f64; // 2100-01-01
-        assert_eq!(safe_timestamp_convert(max_timestamp), 4_102_444_800);
-    }
-
-    #[test]
-    fn test_task_status_from_i32_all_variants() {
-        assert_eq!(TaskStatus::from_i32(0), Some(TaskStatus::Incomplete));
-        assert_eq!(TaskStatus::from_i32(1), None); // unused in real Things 3
-        assert_eq!(TaskStatus::from_i32(2), Some(TaskStatus::Canceled));
-        assert_eq!(TaskStatus::from_i32(3), Some(TaskStatus::Completed));
-        assert_eq!(TaskStatus::from_i32(999), None);
-        assert_eq!(TaskStatus::from_i32(-1), None);
-    }
-
-    #[test]
-    fn test_task_type_from_i32_all_variants() {
-        assert_eq!(TaskType::from_i32(0), Some(TaskType::Todo));
-        assert_eq!(TaskType::from_i32(1), Some(TaskType::Project));
-        assert_eq!(TaskType::from_i32(2), Some(TaskType::Heading));
-        assert_eq!(TaskType::from_i32(3), Some(TaskType::Area));
-        assert_eq!(TaskType::from_i32(999), None);
-        assert_eq!(TaskType::from_i32(-1), None);
-    }
-
-    #[test]
-    fn test_database_pool_config_default_values() {
-        let config = DatabasePoolConfig::default();
-        assert_eq!(config.max_connections, 10);
-        assert_eq!(config.min_connections, 1);
-        assert_eq!(config.connect_timeout, Duration::from_secs(30));
-        assert_eq!(config.idle_timeout, Duration::from_secs(600));
-        assert_eq!(config.max_lifetime, Duration::from_secs(1800));
-        assert!(config.test_before_acquire);
-    }
-
-    #[test]
-    fn test_database_stats_total_items_calculation() {
-        let stats = DatabaseStats {
-            task_count: 10,
-            project_count: 5,
-            area_count: 3,
-        };
-        assert_eq!(stats.total_items(), 18); // 10 + 5 + 3
-
-        // Test with zero values
-        let empty_stats = DatabaseStats {
-            task_count: 0,
-            project_count: 0,
-            area_count: 0,
-        };
-        assert_eq!(empty_stats.total_items(), 0);
-    }
-
-    #[test]
-    fn test_pool_health_status_creation_comprehensive() {
-        let status = PoolHealthStatus {
-            is_healthy: true,
-            pool_size: 8,
-            active_connections: 2,
-            idle_connections: 3,
-            max_connections: 10,
-            min_connections: 1,
-            connection_timeout: Duration::from_secs(30),
-            idle_timeout: Some(Duration::from_secs(600)),
-            max_lifetime: Some(Duration::from_secs(1800)),
-        };
-        assert!(status.is_healthy);
-        assert_eq!(status.pool_size, 8);
-        assert_eq!(status.max_connections, 10);
-    }
-
-    #[test]
-    fn test_pool_metrics_creation_comprehensive() {
-        let metrics = PoolMetrics {
-            pool_size: 8,
-            active_connections: 5,
-            idle_connections: 3,
-            max_connections: 10,
-            min_connections: 1,
-            utilization_percentage: 80.0,
-            is_healthy: true,
-            response_time_ms: 50,
-            connection_timeout: Duration::from_secs(30),
-            idle_timeout: Some(Duration::from_secs(600)),
-            max_lifetime: Some(Duration::from_secs(1800)),
-        };
-        assert_eq!(metrics.pool_size, 8);
-        assert_eq!(metrics.response_time_ms, 50);
-        assert!(metrics.is_healthy);
-    }
-
-    #[test]
-    fn test_comprehensive_health_status_creation_full() {
-        let pool_health = PoolHealthStatus {
-            is_healthy: true,
-            pool_size: 8,
-            active_connections: 2,
-            idle_connections: 3,
-            max_connections: 10,
-            min_connections: 1,
-            connection_timeout: Duration::from_secs(30),
-            idle_timeout: Some(Duration::from_secs(600)),
-            max_lifetime: Some(Duration::from_secs(1800)),
-        };
-
-        let pool_metrics = PoolMetrics {
-            pool_size: 8,
-            active_connections: 5,
-            idle_connections: 3,
-            max_connections: 10,
-            min_connections: 1,
-            utilization_percentage: 80.0,
-            is_healthy: true,
-            response_time_ms: 50,
-            connection_timeout: Duration::from_secs(30),
-            idle_timeout: Some(Duration::from_secs(600)),
-            max_lifetime: Some(Duration::from_secs(1800)),
-        };
-
-        let database_stats = DatabaseStats {
-            task_count: 100,
-            project_count: 20,
-            area_count: 5,
-        };
-
-        let status = ComprehensiveHealthStatus {
-            overall_healthy: true,
-            pool_health,
-            pool_metrics,
-            database_stats,
-            timestamp: Utc::now(),
-        };
-
-        assert!(status.overall_healthy);
-        assert_eq!(status.database_stats.total_items(), 125);
-    }
-
-    #[test]
-    fn test_sqlite_optimizations_default_values() {
-        let opts = SqliteOptimizations::default();
-        assert!(opts.enable_wal_mode);
-        assert!(opts.enable_foreign_keys);
-        assert_eq!(opts.cache_size, -20000);
-        assert_eq!(opts.temp_store, "MEMORY");
-        assert_eq!(opts.mmap_size, 268_435_456);
-        assert_eq!(opts.synchronous_mode, "NORMAL");
-        assert_eq!(opts.journal_mode, "WAL");
-    }
-
-    #[test]
-    fn test_get_default_database_path_format() {
-        let path = get_default_database_path();
-        let path_str = path.to_string_lossy();
-        assert!(path_str.contains("Things Database.thingsdatabase"));
-        assert!(path_str.contains("main.sqlite"));
-        assert!(path_str.contains("Library/Group Containers"));
-    }
-
-    #[test]
-    fn test_discover_things_database_picks_non_default_suffix() {
-        let group_container = TempDir::new().unwrap();
-        let things_dir = group_container.path().join("ThingsData-01AEF");
-        let db_dir = things_dir.join("Things Database.thingsdatabase");
-        std::fs::create_dir_all(&db_dir).unwrap();
-        let db_path = db_dir.join("main.sqlite");
-        std::fs::write(&db_path, b"").unwrap();
-
-        let found = discover_things_database(group_container.path()).unwrap();
-        assert_eq!(found, db_path);
-    }
-
-    #[test]
-    fn test_discover_things_database_prefers_most_recent() {
-        let group_container = TempDir::new().unwrap();
-
-        let make = |suffix: &str| {
-            let dir = group_container
-                .path()
-                .join(format!("ThingsData-{suffix}"))
-                .join("Things Database.thingsdatabase");
-            std::fs::create_dir_all(&dir).unwrap();
-            let db = dir.join("main.sqlite");
-            std::fs::write(&db, b"").unwrap();
-            db
-        };
-
-        let _older = make("OLDER");
-        // 10ms is well above any reasonable filesystem mtime resolution, so
-        // the second file is guaranteed to have a strictly later mtime.
-        std::thread::sleep(std::time::Duration::from_millis(10));
-        let newer = make("NEWER");
-
-        let found = discover_things_database(group_container.path()).unwrap();
-        assert_eq!(found, newer);
-    }
-
-    #[test]
-    fn test_discover_things_database_returns_none_when_empty() {
-        let group_container = TempDir::new().unwrap();
-        assert!(discover_things_database(group_container.path()).is_none());
-    }
-
-    #[test]
-    fn test_discover_things_database_skips_non_matching_dirs() {
-        let group_container = TempDir::new().unwrap();
-        std::fs::create_dir_all(group_container.path().join("SomethingElse")).unwrap();
-        std::fs::create_dir_all(
-            group_container.path().join("ThingsData-EMPTY"), // no main.sqlite inside
-        )
-        .unwrap();
-        assert!(discover_things_database(group_container.path()).is_none());
     }
 
     #[tokio::test]
@@ -4209,194 +3541,6 @@ mod tests {
             health.pool_metrics.is_healthy,
             "Pool metrics should be healthy"
         );
-    }
-
-    // ============================================================================
-    // Date Conversion Tests - Edge Cases
-    // ============================================================================
-
-    #[test]
-    fn test_things_date_negative_returns_none() {
-        // Negative values should return None
-        assert_eq!(things_date_to_naive_date(-1), None);
-        assert_eq!(things_date_to_naive_date(-100), None);
-        assert_eq!(things_date_to_naive_date(i64::MIN), None);
-    }
-
-    #[test]
-    fn test_things_date_zero_returns_none() {
-        // Zero should return None (no date set)
-        assert_eq!(things_date_to_naive_date(0), None);
-    }
-
-    #[test]
-    fn test_things_date_boundary_2001() {
-        use chrono::Datelike;
-        // 1 second after 2001-01-01 00:00:00 should be 2001-01-01
-        let result = things_date_to_naive_date(1);
-        assert!(result.is_some());
-
-        let date = result.unwrap();
-        assert_eq!(date.year(), 2001);
-        assert_eq!(date.month(), 1);
-        assert_eq!(date.day(), 1);
-    }
-
-    #[test]
-    fn test_things_date_one_day() {
-        use chrono::Datelike;
-        // 86400 seconds = 1 day (60 * 60 * 24), should be 2001-01-02
-        let seconds_per_day = 86400i64;
-        let result = things_date_to_naive_date(seconds_per_day);
-        assert!(result.is_some());
-
-        let date = result.unwrap();
-        assert_eq!(date.year(), 2001);
-        assert_eq!(date.month(), 1);
-        assert_eq!(date.day(), 2);
-    }
-
-    #[test]
-    fn test_things_date_one_year() {
-        use chrono::Datelike;
-        // ~365 days should be around 2002-01-01 (365 days * 86400 seconds/day)
-        let seconds_per_year = 365 * 86400i64;
-        let result = things_date_to_naive_date(seconds_per_year);
-        assert!(result.is_some());
-
-        let date = result.unwrap();
-        assert_eq!(date.year(), 2002);
-    }
-
-    #[test]
-    fn test_things_date_current_era() {
-        use chrono::Datelike;
-        // Test a date in the current era (2024)
-        // Days from 2001-01-01 to 2024-01-01 = ~8401 days
-        // Calculation: (2024-2001) * 365 + leap days (2004, 2008, 2012, 2016, 2020) = 23 * 365 + 5 = 8400
-        let days_to_2024 = 8401i64;
-        let seconds_to_2024 = days_to_2024 * 86400;
-
-        let result = things_date_to_naive_date(seconds_to_2024);
-        assert!(result.is_some());
-
-        let date = result.unwrap();
-        assert_eq!(date.year(), 2024);
-    }
-
-    #[test]
-    fn test_things_date_leap_year() {
-        use chrono::{Datelike, TimeZone, Utc};
-        // Test Feb 29, 2004 (leap year)
-        // Days from 2001-01-01 to 2004-02-29
-        let base_date = Utc.with_ymd_and_hms(2001, 1, 1, 0, 0, 0).single().unwrap();
-        let target_date = Utc.with_ymd_and_hms(2004, 2, 29, 0, 0, 0).single().unwrap();
-        let seconds_diff = (target_date - base_date).num_seconds();
-
-        let result = things_date_to_naive_date(seconds_diff);
-        assert!(result.is_some());
-
-        let date = result.unwrap();
-        assert_eq!(date.year(), 2004);
-        assert_eq!(date.month(), 2);
-        assert_eq!(date.day(), 29);
-    }
-
-    // ============================================================================
-    // Timestamp Conversion Tests
-    // ============================================================================
-
-    #[test]
-    fn test_safe_timestamp_convert_normal_values() {
-        // Normal timestamp values should convert correctly
-        let ts = 1_700_000_000.0; // Around 2023
-        let result = safe_timestamp_convert(ts);
-        assert_eq!(result, 1_700_000_000);
-    }
-
-    #[test]
-    fn test_safe_timestamp_convert_zero() {
-        // Zero should return zero
-        assert_eq!(safe_timestamp_convert(0.0), 0);
-    }
-
-    #[test]
-    fn test_safe_timestamp_convert_negative() {
-        // Negative values should return zero (safe fallback)
-        assert_eq!(safe_timestamp_convert(-1.0), 0);
-        assert_eq!(safe_timestamp_convert(-1000.0), 0);
-    }
-
-    #[test]
-    fn test_safe_timestamp_convert_infinity() {
-        // Infinity should return zero (safe fallback)
-        assert_eq!(safe_timestamp_convert(f64::INFINITY), 0);
-        assert_eq!(safe_timestamp_convert(f64::NEG_INFINITY), 0);
-    }
-
-    #[test]
-    fn test_safe_timestamp_convert_nan() {
-        // NaN should return zero (safe fallback)
-        assert_eq!(safe_timestamp_convert(f64::NAN), 0);
-    }
-
-    #[test]
-    fn test_date_roundtrip_known_dates() {
-        use chrono::{Datelike, TimeZone, Utc};
-        // Test roundtrip conversion for known dates
-        // Note: Starting from 2001-01-02 because 2001-01-01 is the base date (0 seconds)
-        // and things_date_to_naive_date returns None for values <= 0
-        let test_cases = vec![
-            (2001, 1, 2), // Start from day 2 since day 1 is the base (0 seconds)
-            (2010, 6, 15),
-            (2020, 12, 31),
-            (2024, 2, 29), // Leap year
-            (2025, 7, 4),
-        ];
-
-        for (year, month, day) in test_cases {
-            let base_date = Utc.with_ymd_and_hms(2001, 1, 1, 0, 0, 0).single().unwrap();
-            let target_date = Utc
-                .with_ymd_and_hms(year, month, day, 0, 0, 0)
-                .single()
-                .unwrap();
-            let seconds = (target_date - base_date).num_seconds();
-
-            let converted = things_date_to_naive_date(seconds);
-            assert!(
-                converted.is_some(),
-                "Failed to convert {}-{:02}-{:02}",
-                year,
-                month,
-                day
-            );
-
-            let result_date = converted.unwrap();
-            assert_eq!(
-                result_date.year(),
-                year,
-                "Year mismatch for {}-{:02}-{:02}",
-                year,
-                month,
-                day
-            );
-            assert_eq!(
-                result_date.month(),
-                month,
-                "Month mismatch for {}-{:02}-{:02}",
-                year,
-                month,
-                day
-            );
-            assert_eq!(
-                result_date.day(),
-                day,
-                "Day mismatch for {}-{:02}-{:02}",
-                year,
-                month,
-                day
-            );
-        }
     }
 
     #[cfg(feature = "advanced-queries")]
